requirements.txt
gitpython==3.1.40
requests==2.31.0
PyYAML==6.0.1
click==8.1.7
tree-sitter==0.24.0
tree-sitter-python==0.23.6
tree-sitter-java==0.23.4
tree-sitter-javascript==0.23.0
tree-sitter-typescript==0.23.0
tree-sitter-cpp==0.23.1
tree-sitter-go==0.23.1
pathspec==0.11.2
python-dotenv==1.0.0
pydantic==2.5.0
dataclasses-json==0.6.1
# Diagram generation dependencies
pillow==10.1.0
matplotlib==3.8.2
networkx==3.2.1
plantuml==2.10.0
lxml==4.9.3
